move lru into the main package
