+++ cache/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "cache.go",
         "ensurespacer.go",
         "http.go",
+        "lru.go",
         "stats.go",
     ],
     importpath = "github.com/buchgr/bazel-remote/cache",
@@ -19,6 +20,7 @@ go_test(
         "cache_test.go",
         "ensurespacer_test.go",
         "http_test.go",
+        "lru_test.go",
         "utils_test.go",
     ],
     embed = [":go_default_library"],
