--- cache/sizedlru/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["sizedlru.go"],
-    importpath = "github.com/buchgr/bazel-remote/cache/sizedlru",
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["sizedlru_test.go"],
-    embed = [":go_default_library"],
-)
